requirements.txt
fastapi==0.109.0
uvicorn[websocket]==0.27.0
websockets==12.0
pydantic==2.5.3
python-dotenv==1.0.0
aiohttp==3.9.1
netifaces==0.11.0